Made the default recipe idempotent
Have made the default.rb recipe idempotent i.e. have ensured that the
Client will first check if the actions listed in the recipe need to be
performed. This will ensure that the cookbook can be run against the
node as many times as required and the actions listed in it will only be
performed when the state of node deviates from what is defined in
the recipe.