New cookbook for LAMP stack
Have added a new cookbook to automate the installation and configuration
of a LAMP stack. Installation and configuration of the web server is
done. MySQL and PHP automation need to be completed. Currently, this
cookbook is being tested on an Ubuntu-16.04 Vagrant box. It was tested
with CentOS-7.2 but caused a few errors, one of which was that the HTTPD
server could not be started or restarted without logging-in to the box
and running 'sudo systemctl start httpd'.